+++ fake-cucumber/javascript/src/SupportCode.ts
@@ -4,7 +4,6 @@ import {
   ParameterTypeRegistry,
 } from '@cucumber/cucumber-expressions'
 import { IdGenerator, messages } from '@cucumber/messages'
-import StackUtils from 'stack-utils'
 import { AnyBody } from './types'
 import ExpressionStepDefinition from './ExpressionStepDefinition'
 import IStepDefinition from './IStepDefinition'
@@ -12,16 +11,6 @@ import IHook from './IHook'
 import Hook from './Hook'
 import IClock from './IClock'
 import { MakeErrorMessage } from './ErrorMessageGenerator'
-
-type RegisterStepDefinition = (
-  expression: string | RegExp,
-  body: AnyBody
-) => void
-
-type RegisterHook = (
-  tagExpressionOrBody: string | AnyBody,
-  body?: AnyBody
-) => void
 
 interface IParameterTypeDefinition {
   name: string
@@ -36,22 +25,6 @@ function defaultTransformer(...args: string[]) {
   return args
 }
 
-function getSourceReference(stackTrace: string): messages.ISourceReference {
-  const stack = new StackUtils({
-    cwd: process.cwd(),
-    internals: StackUtils.nodeInternals(),
-  })
-  const trace = stack.clean(stackTrace)
-  const callSite = stack.parseLine(trace.split('\n')[1])
-  const { file: uri, line } = callSite
-  return new messages.SourceReference({
-    uri,
-    location: new messages.Location({
-      line,
-    }),
-  })
-}
-
 /**
  * This class provides an API for defining step definitions and hooks.
  */
@@ -60,17 +33,7 @@ export default class SupportCode {
   public readonly stepDefinitions: IStepDefinition[] = []
   public readonly beforeHooks: IHook[] = []
   public readonly afterHooks: IHook[] = []
-  public readonly Given = this.registerStepDefinition.bind(
-    this
-  ) as RegisterStepDefinition
-  public readonly When = this.registerStepDefinition.bind(
-    this
-  ) as RegisterStepDefinition
-  public readonly Then = this.registerStepDefinition.bind(
-    this
-  ) as RegisterStepDefinition
-  public readonly Before = this.registerBeforeHook.bind(this) as RegisterHook
-  public readonly After = this.registerAfterHook.bind(this) as RegisterHook
+
   private readonly parameterTypeRegistry = new ParameterTypeRegistry()
   private readonly expressionFactory = new ExpressionFactory(
     this.parameterTypeRegistry
@@ -81,7 +44,9 @@ export default class SupportCode {
     public newId: IdGenerator.NewId,
     public clock: IClock,
     public makeErrorMessage: MakeErrorMessage
-  ) {}
+  ) {
+    console.error('NEW SUPPORT CODE')
+  }
 
   public defineParameterType(
     parameterTypeDefinition: IParameterTypeDefinition
@@ -98,11 +63,11 @@ export default class SupportCode {
     this.parameterTypes.push(parameterType)
   }
 
-  private registerStepDefinition(
+  public registerStepDefinition(
+    sourceReference: messages.ISourceReference,
     expression: string | RegExp,
     body: AnyBody
   ): void {
-    const sourceReference = getSourceReference(new Error().stack)
     try {
       const expr = this.expressionFactory.createExpression(expression)
       const stepDefinition = new ExpressionStepDefinition(
@@ -128,30 +93,34 @@ export default class SupportCode {
     }
   }
 
-  private registerBeforeHook(
+  public registerBeforeHook(
+    sourceReference: messages.ISourceReference,
     tagExpressionOrBody: string | AnyBody,
     body?: AnyBody
   ) {
-    this.beforeHooks.push(this.makeHook(new Error(), tagExpressionOrBody, body))
+    this.beforeHooks.push(
+      this.makeHook(sourceReference, tagExpressionOrBody, body)
+    )
   }
 
-  private registerAfterHook(
+  public registerAfterHook(
+    sourceReference: messages.ISourceReference,
     tagExpressionOrBody: string | AnyBody,
     body?: AnyBody
   ) {
-    this.afterHooks.push(this.makeHook(new Error(), tagExpressionOrBody, body))
+    this.afterHooks.push(
+      this.makeHook(sourceReference, tagExpressionOrBody, body)
+    )
   }
 
   private makeHook(
-    error: Error,
+    sourceReference: messages.ISourceReference,
     tagExpressionOrBody: string | AnyBody,
     body?: AnyBody
   ) {
     const tagExpression =
       typeof tagExpressionOrBody === 'string' ? tagExpressionOrBody : null
     body = typeof tagExpressionOrBody !== 'string' ? tagExpressionOrBody : body
-
-    const sourceReference = getSourceReference(error.stack)
     return new Hook(this.newId(), tagExpression, sourceReference, body)
   }
 }